refactor: move wallet interfaces to types package
Move Wallet, WalletERC20, WalletStableCoin and all StableCoin*
sub-interfaces from wallet/interfaces (interim package) into the
existing types package. The wallet package re-exports them as type
aliases for backward compatibility.
