+++ types/wallet_erc20.go
@@ -1,4 +1,4 @@
-package interfaces
+package types
 
 import (
 	"context"
