+++ wallet/erc20.go
@@ -8,11 +8,10 @@ import (
 	gethTypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/poteto-go/go-alchemy-sdk/constant"
 	"github.com/poteto-go/go-alchemy-sdk/types"
-	walletinterfaces "github.com/poteto-go/go-alchemy-sdk/wallet/interfaces"
 	"golang.org/x/crypto/sha3"
 )
 
-type WalletERC20 = walletinterfaces.WalletERC20
+type WalletERC20 = types.WalletERC20
 
 type walletERC20 struct {
 	w *wallet
