+++ wallet/stablecoin.go
@@ -7,16 +7,16 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	gethTypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/poteto-go/go-alchemy-sdk/constant"
-	walletinterfaces "github.com/poteto-go/go-alchemy-sdk/wallet/interfaces"
+	"github.com/poteto-go/go-alchemy-sdk/types"
 )
 
-type StableCoinMinting = walletinterfaces.StableCoinMinting
-type StableCoinPausing = walletinterfaces.StableCoinPausing
-type StableCoinBlacklisting = walletinterfaces.StableCoinBlacklisting
-type StableCoinMinterAdmin = walletinterfaces.StableCoinMinterAdmin
-type StableCoinRoleAdmin = walletinterfaces.StableCoinRoleAdmin
-type StableCoinInfo = walletinterfaces.StableCoinInfo
-type WalletStableCoin = walletinterfaces.WalletStableCoin
+type StableCoinMinting = types.StableCoinMinting
+type StableCoinPausing = types.StableCoinPausing
+type StableCoinBlacklisting = types.StableCoinBlacklisting
+type StableCoinMinterAdmin = types.StableCoinMinterAdmin
+type StableCoinRoleAdmin = types.StableCoinRoleAdmin
+type StableCoinInfo = types.StableCoinInfo
+type WalletStableCoin = types.WalletStableCoin
 
 type walletStableCoin struct {
 	walletERC20
